+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/atomist/rug/compare/0.26.1...HEAD
+[Unreleased]: https://github.com/atomist/rug/compare/1.0.0-m.1...HEAD
+
+## [1.0.0-m.1] - 2017-04-26
+
+[1.0.0-m.1]: https://github.com/atomist/rug/compare/0.26.1...1.0.0-m.1
 
 ### Added
 
@@ -17,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 -   **BREAKING** TypeScript `Match` now uses property access, not function
 
--   Updated to artifact-source 0.26.0
+-   Updated all Atomist dependencies to 1.0.0-m.1 versions
 
 -   Handlers are now picked up from arrays as well as regular `vars`